Move recording indicator to room layout
