+++ app/src/main/java/com/twilio/video/app/ui/room/ParticipantPrimaryView.kt
@@ -15,14 +15,9 @@
  */
 package com.twilio.video.app.ui.room
 
-import android.animation.ObjectAnimator
-import android.animation.PropertyValuesHolder
-import android.animation.ValueAnimator
 import android.content.Context
 import android.util.AttributeSet
 import android.view.LayoutInflater
-import androidx.core.content.ContextCompat
-import com.twilio.video.app.R
 import com.twilio.video.app.databinding.ParticipantPrimaryViewBinding
 
 internal class ParticipantPrimaryView @JvmOverloads constructor(
@@ -40,7 +35,6 @@ internal class ParticipantPrimaryView @JvmOverloads constructor(
         selectedLayout = binding.selectedLayout
         stubImage = binding.stub
         selectedIdentity = binding.selectedIdentity
-        setupRecordingAnimation()
         setIdentity(identity)
         setState(state)
         setMirror(mirror)
@@ -49,30 +43,5 @@ internal class ParticipantPrimaryView @JvmOverloads constructor(
 
     fun showIdentityBadge(show: Boolean) {
         binding.videoIdentity.visibility = if (show) VISIBLE else GONE
-    }
-
-    fun showRecordingBadge(show: Boolean) {
-        binding.recordingIndicator.apply {
-            visibility = if (show) {
-                VISIBLE
-            } else {
-                GONE
-            }
-        }
-    }
-
-    private fun setupRecordingAnimation() {
-        val recordingDrawable = ContextCompat.getDrawable(context, R.drawable.ic_recording)
-        ObjectAnimator.ofPropertyValuesHolder(recordingDrawable,
-                PropertyValuesHolder.ofInt("alpha", 100, 255)).apply {
-                    target = recordingDrawable
-                    duration = 750
-                    repeatCount = ValueAnimator.INFINITE
-                    repeatMode = ValueAnimator.REVERSE
-                    start()
-                }
-        binding.recordingIndicator.apply {
-            setCompoundDrawablesWithIntrinsicBounds(recordingDrawable, null, null, null)
-        }
     }
 }
