+++ app/src/main/res/layout-v21/room_activity.xml
@@ -33,6 +33,22 @@
 
         </com.google.android.material.appbar.AppBarLayout>
 
+        <com.google.android.material.textview.MaterialTextView
+            android:id="@+id/recording_indicator"
+            android:layout_width="wrap_content"
+            android:layout_height="wrap_content"
+            android:textColor="@android:color/white"
+            android:textSize="14sp"
+            android:text="@string/recording"
+            android:padding="10dp"
+            android:background="@drawable/badge_background"
+            android:drawablePadding="4dp"
+            android:layout_marginTop="100dp"
+            android:layout_marginEnd="16dp"
+            android:visibility="gone"
+            tools:visibility="visible"
+            android:layout_gravity="end" />
+
         <LinearLayout
             android:id="@+id/join_status_layout"
             android:layout_width="match_parent"
