+++ app/src/main/res/layout/participant_primary_view.xml
@@ -69,25 +69,6 @@
             app:layout_constraintStart_toStartOf="parent"
             app:layout_constraintTop_toTopOf="parent"/>
 
-        <com.google.android.material.textview.MaterialTextView
-            android:id="@+id/recording_indicator"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:textColor="@android:color/white"
-            android:textSize="14sp"
-            android:text="@string/recording"
-            android:padding="10dp"
-            android:drawablePadding="4dp"
-            android:layout_marginTop="100dp"
-            android:layout_marginRight="16dp"
-            android:layout_marginEnd="16dp"
-            android:maxLines="1"
-            android:ellipsize="end"
-            android:visibility="gone"
-            tools:visibility="visible"
-            app:layout_constraintEnd_toEndOf="parent"
-            app:layout_constraintTop_toTopOf="parent"/>
-
     </androidx.constraintlayout.widget.ConstraintLayout>
 
 </FrameLayout>
